+++ packages/snap-preact-components/src/components/Organisms/Facet/Facet.stories.tsx
@@ -198,6 +198,26 @@ export default {
 				},
 			},
 		},
+		previewOnFocus: {
+			description: 'Invoke facet value preview upon focus',
+			table: {
+				type: {
+					summary: 'boolean',
+				},
+				defaultValue: { summary: false },
+			},
+			control: { type: 'boolean' },
+		},
+		valueProps: {
+			description: 'Object of facet value props',
+			table: {
+				type: {
+					summary: 'object',
+				},
+				defaultValue: { summary: '{}' },
+			},
+			control: { type: 'object' },
+		},
 		fields: {
 			defaultValue: {},
 			description: 'Change props per facet',